feat(verifier): manual CEV code entry + fix class delete button color
- /verifier index page: type a CEV code to verify a document without scanning
- verifyDocumentByCode() calls /public/verify-code/{tenant}/{code}
- extract AuthenticView/NotRecognizedView to shared result-views.tsx
- class delete button heroAccentBtn -> heroGlassBtn (orange reserved for
  positive focal actions)

Closes #238
